+++ tools/build-package.sh
@@ -80,6 +80,7 @@ find $source_roots -type f \
   ! -path '*/.build-m12/*' \
   ! -path '*/.build-m13/*' \
   ! -path '*/.build-m14/*' \
+  ! -path '*/.build-m15/*' \
   ! -path '*/.libs/*' \
   ! -path '*/.deps/*' \
   -print | LC_ALL=C sort > "$manifest"